fix Warning

+++ pro/lib/SuffixArray.cpp
@@ -8,20 +8,20 @@ using namespace std;
 namespace latte{
     void create_begin_bucket(vector<int>&v,vector<int>&bucket){
         fill(bucket.begin(),bucket.end(),0);
-        for(int i=0;i<v.size();i++)bucket[v[i]]++;
+        for(int i=0;i<(int)v.size();i++)bucket[v[i]]++;
         int sum=0;
-        for(int i=0;i<bucket.size();i++){bucket[i]+=sum;swap(sum,bucket[i]);}
+        for(int i=0;i<(int)bucket.size();i++){bucket[i]+=sum;swap(sum,bucket[i]);}
     }
 
     void create_end_bucket(vector<int>&v,vector<int>&bucket){
         fill(bucket.begin(),bucket.end(),0);
-        for(int i=0;i<v.size();i++)bucket[v[i]]++;
-        for(int i=1;i<bucket.size();i++)bucket[i]+=bucket[i-1];
+        for(int i=0;i<(int)v.size();i++)bucket[v[i]]++;
+        for(int i=1;i<(int)bucket.size();i++)bucket[i]+=bucket[i-1];
     }
 
     void induced_sort(vector<int>&v,vector<int>&sa,int mv,vector<int>&bucket,vector<int>&is_l){
         create_begin_bucket(v,bucket);
-        for(int i=0;i<v.size();i++)if(sa[i]>0&&is_l[sa[i]-1])sa[bucket[v[sa[i]-1]]++]=sa[i]-1;
+        for(int i=0;i<(int)v.size();i++)if(sa[i]>0&&is_l[sa[i]-1])sa[bucket[v[sa[i]-1]]++]=sa[i]-1;
     }
 
     void invert_induced_sort(vector<int>&v,vector<int>&sa,int mv,vector<int>&bucket,vector<int>&is_l){
@@ -40,18 +40,18 @@ namespace latte{
         is_l[v.size()-1]=0;
         for(int i=v.size()-2;i>=0;i--)is_l[i]=v[i]>v[i+1]||(v[i]==v[i+1]&&is_l[i+1]);
         create_end_bucket(v,bucket);
-        for(int i=0;i<v.size();i++)if(is_lms(i))sa[--bucket[v[i]]]=i;
+        for(int i=0;i<(int)v.size();i++)if(is_lms(i))sa[--bucket[v[i]]]=i;
         induced_sort(v,sa,mv,bucket,is_l);
         invert_induced_sort(v,sa,mv,bucket,is_l);
 
         int cur=0;
         vector<int>order(v.size());
-        for(int i=0;i<v.size();i++)if(is_lms(i))order[i]=cur++;
+        for(int i=0;i<(int)v.size();i++)if(is_lms(i))order[i]=cur++;
 
         vector<int>next_v(cur);
         cur=-1;
         int prev=-1;
-        for(int i=0;i<v.size();i++){
+        for(int i=0;i<(int)v.size();i++){
             if(!is_lms(sa[i]))continue;
             bool diff=false;
             for(int d=0;d<v.size();d++){
@@ -66,7 +66,7 @@ namespace latte{
         }
 
         vector<int>re_order(next_v.size());
-        for(int i=0;i<v.size();i++)if(is_lms(i))re_order[order[i]]=i;
+        for(int i=0;i<(int)v.size();i++)if(is_lms(i))re_order[order[i]]=i;
         vector<int>next_sa=sa_is(next_v,cur);
         create_end_bucket(v,bucket);
         for(int i=0;i<sa.size();i++)sa[i]=-1;
@@ -78,7 +78,7 @@ namespace latte{
 
     vector<int> sa_is(string &s){
         vector<int>v(s.size()+1);
-        for(int i=0;i<s.size();i++)v[i]=s[i];
+        for(int i=0;i<(int)s.size();i++)v[i]=s[i];
         return sa_is(v,*max_element(v.begin(),v.end()));
     }
